requirements.txt
asn1crypto==1.5.1
boto3==1.40.5
botocore==1.40.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.0.0
cryptography==45.0.6
filelock==3.18.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
protobuf==5.29.5
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.37.0
sortedcontainers==2.4.0
tomlkit==0.13.3
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
